requirements.txt
asgiref==3.10.0
attrs==25.4.0
charset-normalizer==3.4.4
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
et_xmlfile==2.0.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
openpyxl==3.1.5
pillow==12.0.0
psycopg2-binary==2.9.11
python-dotenv==1.1.1
PyYAML==6.0.3
referencing==0.37.0
reportlab==4.4.4
rpds-py==0.28.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
